requirements.txt

# Optional: dnspython for MX record validation (currently disabled)
# dnspython==2.4.2
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.3.1
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-ratelimit==4.1.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
django-storages==1.14.2
google-cloud-storage==2.13.0
google-cloud-secret-manager==2.16.4
sentry-sdk==1.38.0
django-defender==0.9.7
django-secure==1.0.2
pyopenssl==23.3.0
certifi==2023.11.17
razorpay==1.4.1
paytmchecksum
drf-yasg==1.21.7

# Development dependencies
django-debug-toolbar==4.2.0
django-extensions==3.2.3
pygraphviz==1.11
ipython==7.34.0
pytest-django==4.7.0
coverage==7.3.2
factory-boy==3.3.0
colorama==0.4.6
flake8==6.1.0

# Production dependencies
uvicorn==0.28.0
django-redis==5.4.0
django-prometheus==2.3.1
django-filter==23.5
django-health-check==3.18.1
python-json-logger==2.0.7
django-admin-honeypot==1.1.0
django-clearcache==1.2.1
django-csp==3.7
django-axes==6.1.1
pydantic==2.6.1
httpx==0.26.0
tenacity==8.2.3
boto3==1.34.11  # For AWS S3 storage
cryptography==42.0.0